third_party/WebKit/LayoutTests/svg/text/text-rect-precision-expected.txt

PASS Width of text element is similar to the sum of the width of all characters.
PASS successfullyParsed is true

TEST COMPLETE
Sphinx of black quartz, judge my vow.
.
